+++ native/src/seal/randomgen.cpp
@@ -28,19 +28,25 @@ DWORD last_genrandom_error = 0;
 
 namespace seal
 {
-    uint64_t random_uint64()
+    void random_bytes(seal_byte *buf, size_t count)
     {
-        uint64_t result;
 #if SEAL_SYSTEM == SEAL_SYSTEM_UNIX_LIKE
         random_device rd("/dev/urandom");
-        result = (static_cast<uint64_t>(rd()) << 32) + static_cast<uint64_t>(rd());
+        while (count >= 4)
+        {
+            *reinterpret_cast<uint32_t *>(buf) = rd();
+            buf += 4;
+            count -= 4;
+        }
+        uint32_t last = rd();
+        memcpy(buf, &last, count);
 #elif SEAL_SYSTEM == SEAL_SYSTEM_WINDOWS
         NTSTATUS status = BCryptGenRandom(
             NULL, reinterpret_cast<unsigned char *>(&result), sizeof(result), BCRYPT_USE_SYSTEM_PREFERRED_RNG);
 
         if (BCRYPT_SUCCESS(status))
         {
-            return result;
+            return;
         }
 
         last_bcrypt_error = status;
@@ -49,7 +55,7 @@ namespace seal
         if (!hAdvApi)
         {
             last_genrandom_error = GetLastError();
-            throw runtime_error("Failed to load ADVAPI32.dll");
+            throw runtime_error("Failed to load ADVAPI32.DLL");
         }
 
         BOOLEAN(APIENTRY * RtlGenRandom)
@@ -72,9 +78,15 @@ namespace seal
 #elif SEAL_SYSTEM == SEAL_SYSTEM_OTHER
 #warning "SECURITY WARNING: System detection failed; falling back to a potentially insecure randomness source!"
         random_device rd;
-        result = (static_cast<uint64_t>(rd()) << 32) + static_cast<uint64_t>(rd());
+        while (count >= 4)
+        {
+            *reinterpret_cast<uint32_t *>(buf) = rd();
+            buf += 4;
+            count -= 4;
+        }
+        uint32_t last = rd();
+        memcpy(buf, &last, count);
 #endif
-        return result;
     }
 
     void UniformRandomGeneratorInfo::save_members(ostream &stream) const