+++ CHANGES.md
@@ -5,6 +5,7 @@
 ### Other
 
 - Improved the error message when attempting to configure with `BUILD_SHARED_LIBS=ON` and `SEAL_BUILD_SEAL_C=ON` [(Issue 284)](https://github.com/microsoft/SEAL/issues/284).
+- Added `seal::random_bytes` function in [randomgen.h](native/src/seal/randomgen.h).
 
 ## Version 3.6.2